Rename main directory.
