+++ extensions/ql-vscode/src/model-editor/auto-model-api.ts
