+++ extensions/ql-vscode/src/model-editor/auto-model-codeml-queries.ts
