+++ extensions/ql-vscode/src/model-editor/auto-modeler.ts
