+++ extensions/ql-vscode/src/model-editor/data-extensions-editor-view.ts
