+++ extensions/ql-vscode/src/model-editor/data-schema.json
