+++ extensions/ql-vscode/src/model-editor/extensions-workspace-folder.ts
