+++ extensions/ql-vscode/src/model-editor/external-api-usage.ts
