+++ extensions/ql-vscode/src/model-editor/library.ts
