+++ extensions/ql-vscode/src/model-editor/method-modeling/method-modeling-panel.ts
