+++ extensions/ql-vscode/src/model-editor/method-modeling/method-modeling-view-provider.ts
