+++ extensions/ql-vscode/src/model-editor/methods-usage/methods-usage-data-provider.ts
