+++ extensions/ql-vscode/src/model-editor/modeled-method-fs.ts
