+++ extensions/ql-vscode/src/model-editor/predicates.ts
