+++ extensions/ql-vscode/src/model-editor/queries/java.ts
