+++ extensions/ql-vscode/src/model-editor/queries/query.ts
