+++ extensions/ql-vscode/src/model-editor/shared/hide-modeled-apis.ts
