+++ extensions/ql-vscode/src/model-editor/shared/in-progress-methods.ts
