+++ extensions/ql-vscode/src/model-editor/shared/mode.ts
