+++ extensions/ql-vscode/src/model-editor/shared/modeled-percentage.ts
