+++ extensions/ql-vscode/src/view/data-extensions-editor/DataExtensionsEditor.tsx
@@ -7,19 +7,19 @@ import {
   VSCodeTag,
 } from "@vscode/webview-ui-toolkit/react";
 import { styled } from "styled-components";
-import { ExternalApiUsage } from "../../data-extensions-editor/external-api-usage";
-import { ModeledMethod } from "../../data-extensions-editor/modeled-method";
+import { ExternalApiUsage } from "../../model-editor/external-api-usage";
+import { ModeledMethod } from "../../model-editor/modeled-method";
 import { assertNever } from "../../common/helpers-pure";
 import { vscode } from "../vscode-api";
-import { calculateModeledPercentage } from "../../data-extensions-editor/shared/modeled-percentage";
+import { calculateModeledPercentage } from "../../model-editor/shared/modeled-percentage";
 import { LinkIconButton } from "../variant-analysis/LinkIconButton";
-import { DataExtensionEditorViewState } from "../../data-extensions-editor/shared/view-state";
+import { DataExtensionEditorViewState } from "../../model-editor/shared/view-state";
 import { ModeledMethodsList } from "./ModeledMethodsList";
 import { percentFormatter } from "./formatters";
-import { Mode } from "../../data-extensions-editor/shared/mode";
-import { InProgressMethods } from "../../data-extensions-editor/shared/in-progress-methods";
+import { Mode } from "../../model-editor/shared/mode";
+import { InProgressMethods } from "../../model-editor/shared/in-progress-methods";
 import { getLanguageDisplayName } from "../../common/query-language";
-import { INITIAL_HIDE_MODELED_APIS_VALUE } from "../../data-extensions-editor/shared/hide-modeled-apis";
+import { INITIAL_HIDE_MODELED_APIS_VALUE } from "../../model-editor/shared/hide-modeled-apis";
 
 const LoadingContainer = styled.div`
   text-align: center;
