+++ extensions/ql-vscode/src/view/data-extensions-editor/KindInput.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { ChangeEvent, useCallback, useEffect, useMemo } from "react";
-import type { ModeledMethod } from "../../data-extensions-editor/modeled-method";
+import type { ModeledMethod } from "../../model-editor/modeled-method";
 import { Dropdown } from "../common/Dropdown";
 
 type Props = {
