+++ extensions/ql-vscode/src/view/data-extensions-editor/LibraryRow.tsx
@@ -1,20 +1,20 @@
 import * as React from "react";
 import { useCallback, useMemo, useState } from "react";
 import { styled } from "styled-components";
-import { ExternalApiUsage } from "../../data-extensions-editor/external-api-usage";
-import { ModeledMethod } from "../../data-extensions-editor/modeled-method";
+import { ExternalApiUsage } from "../../model-editor/external-api-usage";
+import { ModeledMethod } from "../../model-editor/modeled-method";
 import { ModeledMethodDataGrid } from "./ModeledMethodDataGrid";
-import { calculateModeledPercentage } from "../../data-extensions-editor/shared/modeled-percentage";
+import { calculateModeledPercentage } from "../../model-editor/shared/modeled-percentage";
 import { percentFormatter } from "./formatters";
 import { Codicon } from "../common";
-import { Mode } from "../../data-extensions-editor/shared/mode";
+import { Mode } from "../../model-editor/shared/mode";
 import {
   VSCodeButton,
   VSCodeDivider,
   VSCodeTag,
 } from "@vscode/webview-ui-toolkit/react";
-import { DataExtensionEditorViewState } from "../../data-extensions-editor/shared/view-state";
-import { InProgressMethods } from "../../data-extensions-editor/shared/in-progress-methods";
+import { DataExtensionEditorViewState } from "../../model-editor/shared/view-state";
+import { InProgressMethods } from "../../model-editor/shared/in-progress-methods";
 
 const LibraryContainer = styled.div`
   background-color: var(--vscode-peekViewResult-background);
